requirements.txt
pygame-ce==2.5.6
pynput==1.8.1
six==1.17.0
